[patch] version bump + fixed get-realtime-breakdown

+++ api_real_time.go
@@ -14,7 +14,7 @@ type RealTimeApiService service
 
 type GetRealtimeBreakdownParams struct {
 	Dimension      string
-	Timestamp      float64
+	Timestamp      int32
 	Filters        []string
 	OrderBy        string
 	OrderDirection string
